Updates `id` commands: now channel and video.
Previously, we had only the `get` command under the `id` group--to get
the channel ID for a given user.  This commit re-names this command into
`getuser`, and adds a new command: `getvideo`, the latter getting the
channel id for a given video.

This commit also updates the grammer to use a [non-personal English
possessive determiner][1].

[1]: http://blog.chrispennello.com/post/140722104245/non-personal-english-possessive-determiner