Improves the `Client.titles` implementation.
This was unused (and apparently somewhat untested) before.  The YouTube
API returns items to us in an order that is potentially different than
the order in which we specified the IDs of the entities in which we were
interested.  This commit fixes the implementation to rectify the
ordering so callers can depend on it.

This commit also adds proper handling for items that are missing from
the API call (e.g., videos marked as private, removed channels, etc.),
including optional exception throwing or "safe" substitution of a
human-readable value indicating an error.

Finally, this commit also updates the documentation of the `titles` call
to reveal that, in fact, you can also pass in `'video'` for the type (it
Just Worked).