Adds title option to `subs ls`.
This commit adds support for fetching the titles of the channels and
playlists to which you are subscribed.  In the mode in which you're
listing out all of the individual videos stored, unsurprisingly, this
will fetch their titles in addition.

This commit removes the `__str__` implementation on subscription objects
in favor of requiring the caller to speficy a particularly-desired
"view".  Currently, these are one that includes IDs and one that
includes titles.

Given that we're usually fetching titles in bulk, and not on a
subscription-by-subscription basis, we take the approach of implementing
support on subscriptions for bulk fetching only.